requirements.txt
requests>=2.28.0
duckduckgo_search>=4.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
